# docs/swagger.yaml
basePath: /api
definitions:
  controllers.audioMessageBody:
    properties:
      base64:
        type: string
      phone:
        type: string
    type: object
  controllers.checkPhonesResponse:
    properties:
      phones:
        items:
          $ref: '#/definitions/controllers.phone'
        type: array
    type: object
  controllers.contactInfoResponse:
    properties:
      info:
        $ref: '#/definitions/services.ContactInfo'
    type: object
  controllers.getMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/services.Message'
        type: array
    type: object
  controllers.getProfileInfoResponse:
    properties:
      info:
        $ref: '#/definitions/services.ContactInfo'
    type: object
  controllers.getQrCodeResponse:
    properties:
      qrCode:
        type: string
    type: object
  controllers.getStatusResponse:
    properties:
      status:
        type: string
    type: object
  controllers.imageMessageBody:
    properties:
      base64:
        type: string
      phone:
        type: string
    type: object
  controllers.phone:
    properties:
      isRegistered:
        type: boolean
      jid:
        properties:
          ad:
            type: boolean
          agent:
            type: integer
          device:
            type: integer
          server:
            type: string
          user:
            type: string
        type: object
      query:
        type: string
    type: object
  controllers.phoneCheckBody:
    properties:
      phones:
        items:
          type: string
        type: array
    type: object
  controllers.sendAudioMessageResponse:
    properties:
      message:
        $ref: '#/definitions/services.Message'
    type: object
  controllers.sendImageMessageResponse:
    properties:
      message:
        $ref: '#/definitions/services.Message'
    type: object
  controllers.sendTextMessageResponse:
    properties:
      message:
        $ref: '#/definitions/services.Message'
    type: object
  controllers.textMessageBody:
    properties:
      phone:
        type: string
      text:
        type: string
    type: object
  services.ContactInfo:
    properties:
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      status:
        type: string
    type: object
  services.Message:
    properties:
      body:
        type: string
      chat:
        type: string
      fromMe:
        type: boolean
      id:
        type: integer
      mediaData:
        properties:
          base64:
            type: string
          mimetype:
            type: string
        type: object
      mediaType:
        type: string
      messageID:
        type: string
      sender:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8900
info:
  contact: {}
  description: API to handle multiple WhatsApp instances
  title: ZapMeow API
  version: "1.0"
paths:
  /{instanceId}/chat/messages:
    post:
      consumes:
      - application/json
      description: Returns chat messages from the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of chat messages
          schema:
            $ref: '#/definitions/controllers.getMessagesResponse'
      summary: Get WhatsApp Chat Messages
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/audio:
    post:
      consumes:
      - application/json
      description: Sends an audio message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Audio message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.audioMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/controllers.sendAudioMessageResponse'
      summary: Send Audio Message on WhatsApp
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/image:
    post:
      consumes:
      - application/json
      description: Sends an image message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Image message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.imageMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/controllers.sendImageMessageResponse'
      summary: Send Image Message on WhatsApp
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/text:
    post:
      consumes:
      - application/json
      description: Sends a text message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Text message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.textMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/controllers.sendTextMessageResponse'
      summary: Send Text Message on WhatsApp
      tags:
      - WhatsApp Chat
  /{instanceId}/check/phones:
    post:
      consumes:
      - application/json
      description: Verifies if the phone numbers in the provided list are registered
        WhatsApp users.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Phone list
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneCheckBody'
      produces:
      - application/json
      responses:
        "200":
          description: List of verified numbers
          schema:
            $ref: '#/definitions/controllers.checkPhonesResponse'
      summary: Check Phones on WhatsApp
      tags:
      - WhatsApp Phone Verification
  /{instanceId}/contact/info:
    get:
      consumes:
      - application/json
      description: Retrieves contact information.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Phone
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Contact Information
          schema:
            $ref: '#/definitions/controllers.contactInfoResponse'
      summary: Get Contact Information
      tags:
      - WhatsApp Contact
  /{instanceId}/logout:
    post:
      consumes:
      - application/json
      description: Logs out from the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
      summary: Logout from WhatsApp
      tags:
      - WhatsApp Logout
  /{instanceId}/profile:
    get:
      consumes:
      - application/json
      description: Retrieves profile information.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile Information
          schema:
            $ref: '#/definitions/controllers.getProfileInfoResponse'
      summary: Get Profile Information
      tags:
      - WhatsApp Profile
  /{instanceId}/qrcode:
    get:
      description: Returns a QR code to initiate WhatsApp login.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: QR code
          schema:
            $ref: '#/definitions/controllers.getQrCodeResponse'
      summary: Get WhatsApp QR Code
      tags:
      - WhatsApp Login
  /{instanceId}/status:
    get:
      consumes:
      - application/json
      description: Returns the status of the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status Response
          schema:
            $ref: '#/definitions/controllers.getStatusResponse'
      summary: Get WhatsApp Instance Status
      tags:
      - WhatsApp Status
swagger: "2.0"